Revert "Use MermaidJS to draw the sequence diagrams in the dynamic credentials documentation"
This reverts commit b1d83afb7950b3dcdcc7824da0838921cbc54956.

We can not rely on a CDN to load the MermaidJS library for the documentation
we package for Tuleap instances. We also can not rely on the Mermaid CLI
since it is not available on the ReadTheDocs.io infrastructure. So without
an ugly hack we can not move those diagrams to Mermaid for now.