Introduce chapter about expected code
It's part of design principles, general orientations as well as stuff that pops
often in code review.

Things were split in several places, I tried to move them all to get something more
consistent.

I'm not convinced about the name of the chapter but I didn't find a better one so here we are.