Add Service special label and description keys
Add a table in Service configuration to list the special Service label
and description keys. Those keys, if left as-is are then translated.
It helps to have this table in case the siteadmin has customised one of
these services' label or description. They can go look that table and
paste the correct key to get translated labels again.