Better structure of user guide
For newcomers (and, honestly, also for old-timers) it is pretty easy to
be lost if the user guide. There are too much top level sections (see
issue #187).

While there is still a lot of work to be done in the individual pages
(some pages are obsolete, others are too big), this commit is here to
group subsections into one. For example "project dashboard", "project
labels", … are now under the project management section.