Suggestion for a better deployment guide structure
This is a follow-up to 105af85c2b46ffddc6cd926210fed77f3d5857a0.

The proposal is to have a more consistent deployment guide. This is done
by moving the following parts to the Deployment guide home page:

- intro section (from Tuleap 11.x page)
- previous releases links (from Tuleap 11.x page)
- list of deprecations (from dedicated page)

Furthermore, now [the installation guide » upgrade to a newer version]
has a link to the deployment guide home page (instead of the Tuleap 11.x).