Add OpenIDConnect provider documentation
This adds documentation aimed at project administrators on how to add an
application to their Tuleap Project and what information they will need
to register Tuleap as an OpenIDConnect provider in the client
application.

It also adds documentation aimed at end-users to explain how to
authorize an application and how to revoke that authorization.

It does NOT explain how to develop an OpenIDConnect client nor details
the OAuth2 flow process.