Move the docs.tuleap.org theme to Webpack 5
While Webpack 5 is still in RC, it should be stable enough to be used on
a non critical project. It also allow us to have a glance of what the
migration to Webpack 5 will look on the main Tuleap repo.

The optimize-css-assets-webpack-plugin plugin has been replaced with the
css-minimizer-webpack-plugin plugin: it already supports Webpack 5
without warning and  it seems to be the most recommended option nowadays
(it is maintained under the webpack-contrib namespace).

The webpack-fix-style-only-entries plugin has been removed has been
removed for now. It throws deprecation warnings when used with Webpack 5
and is not really critical in this repository.