Improve lib documentation for CSS, images and translation
Now that "vite" is used to bundle all libs, it outputs CSS bundle as
"dist/style.css" by default. The doc still assumed its name was customized
like with webpack. We'll rely on that default.
Add an explanation of how to import (and inline) images.
Also explain that the fr_FR.po file must be created manually.