Prevent anchor section links to go under the header on tuleap.org theme
It avoids the annoyance of opening a link and not seeing what the
section is.
Note that the `scroll-margin-top` is always set even when the `.pinned`
class is not set on the `body`. The `.pinned` class is set via a JS
script so the browser is likely to have already scroll to the section
before the class is set (and so the section would end up under the
header).