Add an example of how to attach a file to an Artifact with TUS
There are 3 requests needed to achieve that. The last one especially was
not documented, there was no way to know the shape of the payload
expected to attach a file to an artifact.

Note that I could not test properly this script, I did not set up a
bundler/compiler for that.