Start using Poetry [0] to manage Python deps
This makes possible to load our Python deps directly into a Nix
derivation.
We can now have them directly available into the provided Nix shell
which reduces the required setup that a contributor needs to do to build
the documentation.

It will also allow us to build the tuleap-documentation package as an
"additional package" instead of relying on an independant build step and
2 out of tree Git repositories [1][2].

The `requirements.txt` and `requirements-dev.txt` files are kept for now to
avoid breaking the existing build process.

[0] https://python-poetry.org/
[1] https://tuleap.net/plugins/git/tuleap/deps/tuleap/documentation
[2] https://github.com/Enalean/docker-build-documentation