Flowchart for user registration process
While the code is the only source of truth to understand what is going
on during the registration process, having a flowchart under the eyes
will help to have a broader picture while developing features around the
subject.

Like all documentation, it might become irrelevant some day, that is why
a date is added on top to assert the "freshness" of the graph.

Part of [request #29614][0]: Invited users shouldn't be mandated to re-confirm their email

[0]: https://tuleap.net/plugins/tracker/?aid=29614