Load mermaid script
The initial attempt to use the mermaid CLI to generate the graph during
the build does not work for our production builds.
Running the mermaid CLI involves running Puppeteer. I was not able to
make it work properly inside a container (including with the Chrome sandbox
disabled and all that).

To avoid dealing with `node_modules` and JS build when creating the RPMs
for the local documentation on Tuleap instances, a static mermaid script
is embedded.