Introduce backlog documentation
This replaces as much as possible Agile Dashboard by Backlog.

Redirection is made so that bookmarks to
/user-guide/agile-dashboard.html are redirected to
/user-guide/backlog.html.

Some images have been removed because useless or irrelevant (e.g A.D
homepage)
Some images have been cropped to remove the sidebar so that Agile
Dashboard service does not appear selected for Backlog or Kanban
screenshots.

A notice is added on top of Backlog & Kanban sections to indicates that
Agile Dashboard has been split. This notice should be updated later with
a link to a blog post explaining the change. This notice is planned to
be removed next year.

A second edit is planned in order to find better explanation of Backlog
and to remove mention of "top backlog".

Part of [story #33979][0]: have dedicated Backlog service

[0]: https://tuleap.net/plugins/tracker/?aid=33979