Doc: Artifacts Deletion (#1456)
* Doc: Artifacts Deletion

This commit is here to prepare the page of artifacts deletion. A lot of tickets asked for some explanations. So this commit is here to answer their questions.

* Doc: Artifacts Deletion

I hope that the modification that I put in this commit will suit you.

* Doc: Artifacts Deletion

This commit takes note of the previous review.

* Doc: Artifacts Deletion

This commit is the final one.