Reorganisation of TQL documentation
Information of TQL usage might be confusing.

Since the introduction of TQL documentation, we added new information
alongside new features. But when a user comes to the documentation the
current structure looks like a mess.

For example look at the TOC that is displaying different topics at the
same level:

```
…
+- Tuleap Query Language (TQL)
|  +- TQL on reports
|  +- Queries
|  +- Query construction
|  +- Pro-tips
|  +- Date operators
|  +- TQL on cross-tracker search
|  +- Preconditions for multi-tracker search
|  +- Queries
|  +- Example
|  `- Expert mode
…
```

We have two chapters "Queries" at the same level. There is a chapter
"Export mode" that seems generic but in fact refers only to XTS.

My proposal is to have two different chapters, one for reports, one for
XTS:

```
…
+- Tuleap Query Language (TQL)
|  +- TQL on reports
|  |  +- Queries
|  |  +- Query construction
|  |  `- Pro-tips
|  `- TQL on cross-tracker search
|     +- Preconditions for multi-tracker search
|     +- Queries
|     +- Example
|     `- Expert mode
…
```