docs: Document @ mention in Trackers
Part of story #41090 @ notification in tracker mass-change & artifact
modal

Why?
Changed the heading formats to use simpler characters and be more
explicit on which heading is bigger.
See
https://www.sphinx-doc.org/en/master/usage/restructuredtext/basics.html#sections

Removed the bullet points in the Comments section that were completely
out of proportion, they make more sense as paragraphs, really. Also, I
chose to remove the hard-wrapping as it is annoying to maintain, it
changes nothing to the output and we can adapt our editor windows to the
width we like.

Fixed the English mistakes that I saw.