docs: Remove outdated info from LDAP docs (#1615)
Why?
The LDAP write feature for Tuleap has been removed in June of 2023. We
should remove mention of it in the user docs.
The ability to write plugin configuration in the siteadmin web UI has
also been removed (I don't know when, sometime ago), we should point the
user to the config file (although work is ongoing to replace it).

I changed the headings styles to match Sphinx docs convention. I find
that it avoids breaking my editor's syntax highlighting, like backticks do.

I also switched from "ldap" to "LDAP" because it's the more usual
spelling (it is after all an acronym).

[0] https://tuleap.net/plugins/git/tuleap/tuleap/stable?a=blob&hb=5803f5b0eddfc4fd1eb4662dac7448db225366cc&f=docs%2Fdecisions%2F0024-remove-ldap-write.md
[1] https://www.sphinx-doc.org/en/master/usage/restructuredtext/basics.html#sections