feat: Tuleap docs rebranding (#1726)
part of story #45211: Tuleap 17.0 new theme

How to test:
- Rebuild the CSS with the following command:
$ npm i && npm run build
- Rebuild the HTML with the following command:
$ make watch-html-tuleap-org
- Inspect the docs
- How-to guides & User guides use the new orange theme color (links and
header have this color)
- The logo image on the top-left is updated.
- The favicon is updated (you can check the various versions with the
"Favicon Detector" extension for Firefox).
- The illustrations on the front page use the new logo.
- The font is updated to "Public Sans Variable".

Why?
We do not have a satisfactory way to cache-bust assets like CSS and JS.
I used manual renaming ("style" -> "style-2") to avoid cache issues, but
I know it's a broken process. It's nonetheless better than nothing.

I did not bother with webmanifest images, the favicons should be
sufficient.

Only "woff2" format is included for Public Sans font. woff2 is
considered Baseline since about 2018, it was only IE11 that kept us
using TTF and woff format.

I used SVG files for front page illustrations so they can be cached. The
only exception was the developer guide image, because it uses a CSS
class to change the gradient colors, so it must be inline. We could also
"hardcode" the color directly to the SVG "stop-color" attributes and
make it an external file, I guess the CSS class is more maintainable.

Co-authored-by: Nicolas Terray <nicolas.terray@enalean.com>