Use the new dark color
We should use the new dark brown color for text color (and for few
elements that has a dark background).

Instead of using the variable `$grey` we now use `$dark-color` and
`$typo-default-text-color` to reuse the same terminology as in Tuleap
source code. This brings some developer consistency.

Part of [story #45211][0]: Tuleap 17.0 new theme

[0]: https://tuleap.net/plugins/tracker/?aid=45211